+++ app/scripts/components/analysis/define/page-hero-actions.tsx
@@ -1,7 +1,8 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useMemo } from 'react';
 import { Link } from 'react-router-dom';
 import { sticky } from 'tippy.js';
 import { Feature, MultiPolygon, Polygon } from 'geojson';
+// import styled from 'styled-components';
 import { Button, ButtonProps } from '@devseed-ui/button';
 import {
   CollecticonTickSmall,
@@ -20,15 +21,17 @@ import {
 import { DatasetLayer } from 'veda/thematics';
 
 import { analysisParams2QueryString } from '../results/use-analysis-params';
+
+import useSavedSettings from './use-saved-settings';
+
 import { Tip } from '$components/common/tip';
 import { resourceNotFound } from '$components/uhoh';
 import { thematicAnalysisPath } from '$utils/routes';
 import { useThematicArea } from '$utils/thematics';
 import { composeVisuallyDisabled } from '$utils/utils';
 import { useMediaQuery } from '$utils/use-media-query';
-import DropMenuItemButton from '$styles/drop-menu-item-button';
-import { getDateRangeFormatted } from '../utils';
-import styled from 'styled-components';
+
+// import DropMenuItemButton from '$styles/drop-menu-item-button';
 
 const SaveButton = composeVisuallyDisabled(Button);
 
@@ -41,13 +44,6 @@ interface PageHeroActionsProps {
   datasetsLayers?: DatasetLayer[];
   aoi?: Feature<Polygon> | null;
 }
-interface SavedSettings {
-  url: string;
-  label: string;
-}
-
-const SAVED_SETTINGS_KEY = 'analysisSavedSettings';
-const MAX_SAVED_SETTINGS = 5;
 
 export default function PageHeroActions({
   size,
@@ -83,44 +79,13 @@ export default function PageHeroActions({
     });
   }, [start, end, datasetsLayers, aoi]);
 
-  const onGenerateClick = useCallback(() => {
-    console.log(datasetsLayers);
-    const savedSettingsRaw = localStorage.getItem(SAVED_SETTINGS_KEY);
-    try {
-      let savedSettings: SavedSettings[] = savedSettingsRaw
-        ? JSON.parse(savedSettingsRaw)
-        : [];
-      if (!savedSettings.find((s) => s.url === analysisParamsQs)) {
-        savedSettings = [
-          {
-            url: analysisParamsQs,
-            label: `${datasetsLayers
-              ?.map((dL) => dL.name)
-              .join(', ')} - ${getDateRangeFormatted(start, end)}`
-          },
-          ...savedSettings,
-        ];
-        if (savedSettings.length > MAX_SAVED_SETTINGS) {
-          savedSettings = savedSettings.slice(1);
-        }
-        localStorage.setItem(SAVED_SETTINGS_KEY, JSON.stringify(savedSettings));
-      }
-    } catch (e) {}
-  }, [analysisParamsQs]);
-
-  // Only need to read localStorage at component mount, because whenever the localStorage item is updated,
-  // this components gets unmounted anyways (navigating from the page using the 'Generate' button)
-  const [savedSettingsList, setSavedSettingsList] = useState<SavedSettings[]>(
-    []
+  const { onGenerateClick, thematicAreaSavedSettingsList } = useSavedSettings(
+    thematic.data.id,
+    analysisParamsQs,
+    start,
+    end,
+    datasetsLayers
   );
-  useEffect(() => {
-    const savedSettingsRaw = localStorage.getItem('analysisSavedSettings');
-    try {
-      if (savedSettingsRaw) {
-        setSavedSettingsList(JSON.parse(savedSettingsRaw));
-      }
-    } catch (e) {}
-  }, []);
 
   let tipContents;
 
@@ -181,7 +146,7 @@ export default function PageHeroActions({
           <CollecticonTickSmall /> Generate
         </Button>
       )}
-      {savedSettingsList.length > 0 && (
+      {thematicAreaSavedSettingsList.length > 0 && (
         <Toolbar size='small'>
           <Dropdown
             alignment='right'
@@ -198,8 +163,8 @@ export default function PageHeroActions({
           >
             <DropTitle>Retrieve previous settings</DropTitle>
             <DropMenu>
-              {savedSettingsList.map((savedSettings) => (
-                <li>
+              {thematicAreaSavedSettingsList.map((savedSettings) => (
+                <li key={savedSettings.url}>
                   <DropMenuItem data-dropdown='click.close'>
                     <Button
                       forwardedAs={Link}
